refactor: view/domain 역할 분리
