+++ src/main/java/game/core/domain/Game.java
@@ -1,4 +1,4 @@
-package game;
+package game.core.domain;
 
 /**
  * @author : yusik
