+++ src/main/java/game/core/domain/ResultEntity.java
@@ -1,4 +1,4 @@
-package game;
+package game.core.domain;
 
 /**
  * @author : yusik
