+++ src/test/java/game/core/view/InputViewTest.java
@@ -1,4 +1,4 @@
-package game;
+package game.core.view;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
